individual capacity in table + floatization

+++ src/estimation.py
@@ -3,8 +3,8 @@ from member import MemberList
 
 class Estimation(BaseModel):
 
-    sprint_days: int
-    scrum_factor: int
+    sprint_days: float
+    scrum_factor: float
     capacity: float
 
     member_list: MemberList
